fix: resolve a bug that occurs while updating a link in the table mode view
