+++ antarest/study/business/model/link_model.py
@@ -12,19 +12,122 @@
 import typing as t
 
 from antares.study.version import StudyVersion
-from pydantic import ConfigDict, Field, model_validator
+from pydantic import BeforeValidator, ConfigDict, Field, PlainSerializer, model_validator
 
 from antarest.core.exceptions import LinkValidationError
 from antarest.core.serialization import AntaresBaseModel
 from antarest.core.utils.string import to_camel_case, to_kebab_case
+from antarest.study.business.enum_ignore_case import EnumIgnoreCase
 from antarest.study.model import STUDY_VERSION_8_2
-from antarest.study.storage.rawstudy.model.filesystem.config.links import (
-    AssetType,
-    FilterOption,
-    LinkStyle,
-    TransmissionCapacity,
-    comma_separated_enum_list,
-)
+
+
+class AssetType(EnumIgnoreCase):
+    """
+    Enum representing the type of asset for a link between two areas.
+
+    Attributes:
+        AC: Represents an Alternating Current link. This is the most common type of electricity transmission.
+        DC: Represents a Direct Current link. This is typically used for long-distance transmission.
+        GAZ: Represents a gas link. This is used when the link is related to gas transmission.
+        VIRT: Represents a virtual link. This is used when the link doesn't physically exist
+            but is used for modeling purposes.
+        OTHER: Represents any other type of link that doesn't fall into the above categories.
+    """
+
+    AC = "ac"
+    DC = "dc"
+    GAZ = "gaz"
+    VIRT = "virt"
+    OTHER = "other"
+
+
+class TransmissionCapacity(EnumIgnoreCase):
+    """
+    Enum representing the transmission capacity of a link.
+
+    Attributes:
+        INFINITE: Represents a link with infinite transmission capacity.
+            This means there are no limits on the amount of electricity that can be transmitted.
+        IGNORE: Represents a link where the transmission capacity is ignored.
+            This means the capacity is not considered during simulations.
+        ENABLED: Represents a link with a specific transmission capacity.
+            This means the capacity is considered in the model and has a certain limit.
+    """
+
+    INFINITE = "infinite"
+    IGNORE = "ignore"
+    ENABLED = "enabled"
+
+
+class LinkStyle(EnumIgnoreCase):
+    """
+    Enum representing the style of a link in a network visualization.
+
+    Attributes:
+        DOT: Represents a dotted line style.
+        PLAIN: Represents a solid line style.
+        DASH: Represents a dashed line style.
+        DOT_DASH: Represents a line style with alternating dots and dashes.
+    """
+
+    DOT = "dot"
+    PLAIN = "plain"
+    DASH = "dash"
+    DOT_DASH = "dotdash"
+    OTHER = "other"
+
+
+class FilterOption(EnumIgnoreCase):
+    """
+    Enum representing the time filter options for data visualization or analysis in Antares Web.
+
+    Attributes:
+        HOURLY: Represents filtering data by the hour.
+        DAILY: Represents filtering data by the day.
+        WEEKLY: Represents filtering data by the week.
+        MONTHLY: Represents filtering data by the month.
+        ANNUAL: Represents filtering data by the year.
+    """
+
+    HOURLY = "hourly"
+    DAILY = "daily"
+    WEEKLY = "weekly"
+    MONTHLY = "monthly"
+    ANNUAL = "annual"
+
+
+def validate_filters(
+    filter_value: t.Union[t.List[FilterOption], str], enum_cls: t.Type[FilterOption]
+) -> t.List[FilterOption]:
+    if isinstance(filter_value, str):
+        if not filter_value.strip():
+            return []
+
+        filter_accepted_values = [e for e in enum_cls]
+
+        options = filter_value.replace(" ", "").split(",")
+
+        invalid_options = [opt for opt in options if opt not in filter_accepted_values]
+        if invalid_options:
+            raise LinkValidationError(
+                f"Invalid value(s) in filters: {', '.join(invalid_options)}. "
+                f"Allowed values are: {', '.join(filter_accepted_values)}."
+            )
+        options_enum: t.List[FilterOption] = list(dict.fromkeys(enum_cls(opt) for opt in options))
+        return options_enum
+
+    return filter_value
+
+
+def join_with_comma(values: t.List[FilterOption]) -> str:
+    return ", ".join(value.name.lower() for value in values)
+
+
+comma_separated_enum_list = t.Annotated[
+    t.List[FilterOption],
+    BeforeValidator(lambda x: validate_filters(x, FilterOption)),
+    PlainSerializer(lambda x: join_with_comma(x)),
+]
 
 DEFAULT_COLOR = 112
 FILTER_VALUES: t.List[FilterOption] = [
