+++ tests/integration/study_data_blueprint/test_link.py
@@ -9,12 +9,10 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
-from sys import stderr
-
 import pytest
 from starlette.testclient import TestClient
 
-from antarest.study.storage.rawstudy.model.filesystem.config.links import TransmissionCapacity
+from antarest.study.business.model.link_model import TransmissionCapacity
 from tests.integration.prepare_proxy import PreparerProxy
 
 
@@ -276,6 +274,34 @@ class TestLink:
             "colorr": 112,
             "displayComments": True,
             "filterSynthesis": "",
+            "filterYearByYear": "hourly, daily, weekly, monthly, annual",
+            "hurdlesCost": False,
+            "linkStyle": "plain",
+            "linkWidth": 1.0,
+            "loopFlow": False,
+            "transmissionCapacities": "enabled",
+            "usePhaseShifter": False,
+        }
+        assert expected == res.json()
+
+        # Test create link with double value in filter
+
+        client.delete(f"/v1/studies/{study_id}/links/{area1_id}/{area2_id}")
+        res = client.post(
+            f"/v1/studies/{study_id}/links",
+            json={"area1": area1_id, "area2": area2_id, "filterSynthesis": "hourly, hourly"},
+        )
+
+        assert res.status_code == 200, res.json()
+        expected = {
+            "area1": "area 1",
+            "area2": "area 2",
+            "assetType": "ac",
+            "colorb": 112,
+            "colorg": 112,
+            "colorr": 112,
+            "displayComments": True,
+            "filterSynthesis": "hourly",
             "filterYearByYear": "hourly, daily, weekly, monthly, annual",
             "hurdlesCost": False,
             "linkStyle": "plain",
